Added a demo: "Stack"
Using the sticky stack will give you super powers. Well, not literally
of course, but it could be a handy functionality indeed. Stacks are a
way to have multiple independent sticky elements working together. See
for yourself and scroll this page. As soon as multiple sticky elements
become sticky the others neatly adjust their offset automatically.

It could be useful if you have multiple ui element on your page that are
on different places but still need to be always shown. And thanks to the
service you can also get the current stack height or the height of the
stack at a certain scroll position. This can be useful if you want to
scroll to a certain position on the page but you're not sure how high
your sticky element stack will be.