Added demo "Container"
Container are useful for constraining a sticky element. Setting a
container is done by adding a container attribute the id of an element.
It simply checks the bounding box of that container element. If the
container's bounding box is not big enough to contain the sticky element
it starts to disappear.